change type of `pan_angle` and `tilt_angle` of `PointOfMotion` and `CameraState` from `int` to `float`
If the gimbal is not at the starting point, `_previous_point` of
`BaseCamPathOfMotionCameraController` is set to a `PointOfMotion` with the
current angles of the gimbal. The angles have been rounded/converted to `int`,
since `PointOfMotion` has represented angles in `int`, but since they are
degrees, they could and should be `float`.
`_is_current_point_reached` compares angles of the gimbal with the target,
which may be the previous angles of the gimbal. If they would be rounded,
the result is unnecessary imprecise.

Note: `GetAnglesInCmd` represents angles (radians) in `int`,
but angles are converted to degrees of type `float`.