fix: speeds are reset based on the wrong points
The naming of the variables was misleading.
Actually, `*_clockwise` of the previous point has been compared with the
pre-previous point. `self.next_point()` has to be called before
`self.current_point()`.
The gimbal should move to the next point. It was moving to the point before
the next point. If the gimbal should keep moving in the same direction,
the direction (`*_clockwise`) of the previous target and the new target are
the same. Otherwise, the speed has to be reset. Anyway, the pre-previous point
is not relevant anymore.