fix: use actual image size to calculate speed based on distance in image
The magic number 320 (half of the default live view image width) had been used.
However, the used live view size might be changed using a CLI argument
and the height should be used for the calculation of the tilt speed.
Further, the speed should not been rounded to an int,
since pan and tilt speeds are floats instead of ints in the meantime.